requirements.txt
# Core Django
Django==5.1.15
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Static files and middleware
whitenoise==6.6.0

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0

# API and CORS
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Cache and sessions
redis==5.0.1

# Development tools (optional for production)
django-extensions==3.2.3

# Email
django-environ==0.11.2

# Monitoring (optional)
sentry-sdk==1.38.0
